requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
flask-cors==4.0.0
pymongo==4.6.1
pyjwt==2.8.0
werkzeug==3.0.1